代码优化添加实体类转换工具

+++ rain-resource/src/main/java/com/rao/service/resource/ResourceLocationsConfigService.java
@@ -23,5 +23,5 @@ public interface ResourceLocationsConfigService {
      * 查询记录数
      * @return
      */
-    int count();
+    Long count();
 }
